backend/requirements.txt
# Core FastAPI and dependencies
fastapi
uvicorn[standard]

# Data validation and settings
pydantic
pydantic-settings

# HTTP client for external APIs
httpx

# Authentication
python-jose[cryptography]
python-multipart

# Database (Supabase PostgreSQL)
supabase  # Supabase Python client

# Password hashing
passlib[bcrypt]

# Environment variables
python-dotenv

# Testing
pytest
pytest-asyncio

# AI integration
openai==1.55.3

# Development tools
pylint
bandit